lambda/requirements.txt
pylint
python-dotenv

